+++ src/core/RTransaction.cpp
@@ -1164,7 +1164,7 @@ QList<RPropertyChange> RTransaction::getPropertyChanges(RObject::Id id) const {
  * Stream operator for QDebug
  */
 QDebug operator<<(QDebug dbg, RTransaction& t) {
-    dbg.nospace() << "RTransaction(" << QString("%1").arg((long)&t, 0, 16);
+    dbg.nospace() << "RTransaction(" << QString("%1").arg((unsigned long long int)&t, 0, 16);
 
     dbg.nospace() << ", id: " << t.getId();
     dbg.nospace() << ", types: " << t.getTypes();
